Remove config from Bespoke::Base->init.
When I started coding I was thinking about being able to pass around
different config objects and modify behavior on the fly using that for
things like testing. This turned out to be a dumb and ugly idea.

This change removes config from the return from init() and switches
users of it over to $self->get_config(@_).  I'm inclined to ditch
that soon too since it's not all that useful and just go with the
singleton + its fancy import().

It can wait though.